+++ src/shared/constants/shortcuts.ts
@@ -29,7 +29,6 @@ export const ShortcutsMap = {
     accelerator: "CmdOrCtrl+F",
   },
 
-  // todo: move as new window
   "ui:open-assistant-panel": {
     channel: "shortcut:ui:open-assistant-panel",
     label: "AI Assistant",